# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movieflix API",
        "description": "API responsável por gerenciar os filmes do Movieflix",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Listar todos os filmes",
                "responses": {
                    "200": {
                        "description": "Lista de filmes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movie"
                ],
                "summary": "Adicionar um novo filme",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Filme criado com sucesso",
                        "content": {
                            "application/json": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Atualizar um filme por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Filme atualizado com sucesso!"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "505": {
                        "description": "Falha ao atualizar o filme"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie"
                ],
                "summary": "Deletar um filme por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Filme deletado com sucesso"
                    },
                    "404": {
                        "description": "Filme não encontrado"
                    },
                    "505": {
                        "description": "Falha ao deletar o filme"
                    }
                }
            }
        },
        "/movies/{genreName}": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Listar filmes por gênero",
                "parameters": [
                    {
                        "name": "genreName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de filmes por gênero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Falha ao filtrar filmes por gênero"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "string",
                        "description": "Gênero do filme"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    },
                    "genres": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID do gênero"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome do gênero"
                            }
                        }
                    },
                    "language_id": {
                        "type": "string",
                        "example": "Idioma do filme"
                    },
                    "languages": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "description": "ID da língua"
                            },
                            "name": {
                                "type": "string",
                                "description": "Nome da língua"
                            }
                        }
                    },
                    "oscar_count": {
                        "type": "integer",
                        "example": "Quantidade de Oscars que o filme ganhou"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "genre_id",
                    "releaseDate",
                    "language_id",
                    "oscar_count"
                ]
            },
            "MovieInput": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID da lígua do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de oscars que o filme ganhou"
                    }
                },
                "required": [
                    "title",
                    "genre_id",
                    "releaseDate",
                    "language_id",
                    "oscar_count"
                ]
            }
        },
        "responses": {
            "MovieResponse": {
                "$ref": "#/components/schemas/Movie"
            },
            "NotFoundResponse": {
                "description": "Filme não encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Filme não encontrado"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}